+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.95
 
 - Crunchyroll: Removed header adjustment of big player style, because fullscreen button not clickable.
-- Crunchyroll: fixed skip intro.
+- Crunchyroll: fixed skip intro, fixed removed selected speed when changing episode.
 
 
 ## 1.1.94